propertySet included and optional baseline
each region can get a property set - the properties are inherited to the children but can be deleted (for example <property key="lang" value=""/>)
coords does not have to have an element Coords. The idea is that one can store references (on line- or block layer) in the xml structure without having polygons. Later a Text2Image-Tools can link found polygons to the references.